+++ backend/modules/content/controllers/ProjectController.php
@@ -2,6 +2,7 @@
 
 namespace backend\modules\content\controllers;
 
+use common\models\Story;
 use Yii;
 use yii\base\Model;
 use common\models\Project;
@@ -86,7 +87,7 @@ class ProjectController extends Controller
         }
 
         return $this->render('create', [
-            'model' => $model,
+            'model' => $model
         ]);
     }
 
@@ -100,7 +101,7 @@ class ProjectController extends Controller
     public function actionUpdate($id)
     {
         $model = $this->findModel($id);
-
+    //    $story = new Story();
 
         if ($model->load(Yii::$app->request->post())) {
             if ($model->validate()) {
@@ -113,7 +114,11 @@ class ProjectController extends Controller
                     $model->main_photo = $path;
                     $file->saveAs('uploads/project/' . $path);
                     if (isset($old_name)) {
-                        unlink('uploads/project/' . $old_name);
+                        if($old_name==''){
+
+                        }else {
+                            unlink('uploads/project/' . $old_name);
+                        }
                     } else {
                         // Do nothing
                     }
@@ -128,6 +133,7 @@ class ProjectController extends Controller
 
         return $this->render('update', [
             'model' => $model,
+      //      'story' => $story
             ]);
     }
